Rename Prometheus collector files to enterprise billing collector files
